+++ packages/core/src/types/curi.ts
@@ -40,6 +40,7 @@ export interface RouterOptions {
   cache?: Cache;
   pathnameOptions?: PathFunctionOptions;
   emitRedirects?: boolean;
+  sync?: boolean;
 }
 
 export interface CurrentResponse {