+++ packages/core/src/types/route.ts
@@ -1,8 +1,8 @@
 import { RegExpOptions, Key } from "path-to-regexp";
 
-import { Params } from "./response";
-import { Addons } from "./addon";
 import { LocationDetails } from "@hickory/root";
+import { Params, ResponseProps } from './response';
+import { Addons } from './addon';
 
 export type ParamParser = (input: string) => any;
 export interface ParamParsers {
@@ -86,4 +86,9 @@ export interface InternalRoute {
 export interface Match {
   route: InternalRoute;
   params: Params;
+}
+
+export interface MatchedRoute {
+  route: InternalRoute;
+  props: ResponseProps;
 }