+++ packages/core/types/types/curi.d.ts
@@ -32,6 +32,7 @@ export interface RouterOptions {
     cache?: Cache;
     pathnameOptions?: PathFunctionOptions;
     emitRedirects?: boolean;
+    sync?: boolean;
 }
 export interface CurrentResponse {
     response: Response;