+++ packages/core/types/types/route.d.ts
@@ -1,7 +1,7 @@
 import { RegExpOptions, Key } from "path-to-regexp";
-import { Params } from "./response";
-import { Addons } from "./addon";
 import { LocationDetails } from "@hickory/root";
+import { Params, ResponseProps } from './response';
+import { Addons } from './addon';
 export declare type ParamParser = (input: string) => any;
 export interface ParamParsers {
     [key: string]: ParamParser;
@@ -73,4 +73,8 @@ export interface InternalRoute {
 export interface Match {
     route: InternalRoute;
     params: Params;
+}
+export interface MatchedRoute {
+    route: InternalRoute;
+    props: ResponseProps;
 }